# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cadastro de Livros API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/livro/create": {
            "post": {
                "tags": [
                    "Livro"
                ],
                "summary": "Cadastra os dados do Livro",
                "description": "Cadastra os dados do Livro",
                "operationId": "76986317a04038c327af4bb9da625838",
                "parameters": [
                    {
                        "name": "Titulo",
                        "in": "query",
                        "description": "Insira o título do livro",
                        "required": true
                    },
                    {
                        "name": "Editora",
                        "in": "query",
                        "description": "Insira a editora do livro",
                        "required": true
                    },
                    {
                        "name": "Edicao",
                        "in": "query",
                        "description": "Insira a edição do livro",
                        "required": true
                    },
                    {
                        "name": "AnoPublicacao",
                        "in": "query",
                        "description": "Insira o ano de publicação do livro",
                        "required": true
                    },
                    {
                        "name": "Autor_CodAu",
                        "in": "query",
                        "description": "Insira o autor do livro",
                        "required": true
                    },
                    {
                        "name": "Assunto_CodAs",
                        "in": "query",
                        "description": "Insira o assunto do livro",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/livro/edit/{id}": {
            "get": {
                "tags": [
                    "Livro"
                ],
                "summary": "Retorna dados específicos de um Livro",
                "description": "Retorna dados específicos de um Livro",
                "operationId": "363d22f94e4abd7dc2ff964310f7f6a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do livro",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/livro/update/{id}": {
            "put": {
                "tags": [
                    "Livro"
                ],
                "summary": "Atualiza dados de um Livro",
                "description": "Atualiza dados de um Livro",
                "operationId": "dcf416ebc59f1f4bbb24ae3cfda59f1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do livro",
                        "required": true
                    },
                    {
                        "name": "Titulo",
                        "in": "query",
                        "description": "Insira o título do livro",
                        "required": true
                    },
                    {
                        "name": "Editora",
                        "in": "query",
                        "description": "Insira a editora do livro",
                        "required": true
                    },
                    {
                        "name": "Edicao",
                        "in": "query",
                        "description": "Insira a edição do livro",
                        "required": true
                    },
                    {
                        "name": "AnoPublicacao",
                        "in": "query",
                        "description": "Insira o ano de publicação do livro",
                        "required": true
                    },
                    {
                        "name": "Autor_CodAu",
                        "in": "query",
                        "description": "Insira o autor do livro",
                        "required": true
                    },
                    {
                        "name": "Assunto_CodAs",
                        "in": "query",
                        "description": "Insira o assunto do livro",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/livro/delete/{id}": {
            "delete": {
                "tags": [
                    "Livro"
                ],
                "summary": "Deleta um Livro específico",
                "description": "Deleta um Livro específico",
                "operationId": "0eefb79331d1cf8bddcf885e7765e4c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do livro",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}